Some Improvements

+++ functionality/trainee.py
@@ -1,6 +1,6 @@
 import re
-from flask_restful import Resource, abort,reqparse
-from db import mycursor,mydb
+from flask_restful import Resource, abort, reqparse
+from db import mydb, mycursor, id_exist
 from args_parser import *
 
 # Trainee
@@ -9,43 +9,40 @@ from args_parser import *
 
 class Trainee(Resource):
     def get(self, trainee_id):
-        sql = 'SELECT * FROM TraineeList WHERE ID= {0}'.format(trainee_id)
-        mycursor.execute(sql)
-        result = mycursor.fetchall()
         # Check if exists
-        if len(result) == 0:
-            abort(404, message="Trainee with the specified ID {} doesn't exist".format(
-                trainee_id))
-        else:
+        if id_exist(trainee_id):
+
+            sql = 'SELECT * FROM TraineeList WHERE ID= {0}'.format(trainee_id)
+            mycursor.execute(sql)
+            result = mycursor.fetchall()
+            mydb.commit()
             # JSONify output
             result = [dict(zip([key[0] for key in mycursor.description], row))
                       for row in result]
-        return result
+        else:
+            abort(404, message="Trainee with the specified ID {} doesn't exist".format(
+                trainee_id))
+        return result, 200
 
     def delete(self, trainee_id):
-        sql = '{0} {1}'.format(
-            "DELETE FROM TraineeList WHERE ID =", trainee_id)
-        mycursor.execute(sql)
-        mydb.commit()
-        # fix AUTO_INCREMENT value
-        sql = "SELECT MAX(ID) FROM TraineeList"
-        mycursor.execute(sql)
-        max_id = str(mycursor.fetchall()[0])
-        max_id = re.sub('[(),]', '', max_id)
-        sql = '{0} {1}'.format(
-            "ALTER TABLE TraineeList AUTO_INCREMENT = ", max_id)
-        mycursor.execute(sql)
-        return '{0} {1}'.format("1", "record(s) deleted"), 200
+        if id_exist(trainee_id):
+            sql = '{0} {1}'.format(
+                "DELETE FROM TraineeList WHERE ID =", trainee_id)
+            mycursor.execute(sql)
+            mydb.commit()
+        else:
+            abort(404, message="Trainee with the specified ID {} doesn't exist".format(
+                trainee_id))
+        return '{0} {1}'.format("1 record deleted"), 200
 
     def put(self, trainee_id):
         args = parser.parse_args()
         # update only if arg value is not null
-        Trainee.get(self, trainee_id)
+
         for param in args:
             if args[param] is not None:
                 sql = 'UPDATE TraineeList SET {0} = "'"{1}"'" WHERE ID={2}'.format(
                     param, args[param], trainee_id)
                 mycursor.execute(sql)
                 mydb.commit()
-        print(mycursor.rowcount, "record(s) affected")
-        return 201
+        return "1 record affected", 201
